requirements.txt
yfinance>=0.2.40
pyyaml==6.0.1
requests>=2.31.0
flask>=2.3.0
pandas>=1.3.0
lxml>=4.9.0
streamlit>=1.28.0
